requirements.txt
asgiref==3.11.0
attrs==25.4.0
channels==4.3.2
channels_redis==4.3.0
Django==6.0
django-filter==25.2
django-redis==6.0.0
django-seed==0.3.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
drf-spectacular-sidecar==2026.1.1
drf-yasg==1.21.11
Faker==40.1.2
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
msgpack==1.1.2
packaging==25.0
pillow==12.0.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
pytz==2025.2
PyYAML==6.0.3
redis==7.1.0
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.5
toposort==1.10
typing_extensions==4.15.0
uritemplate==4.2.0
